Fixed bug in GroupLogic; the hash of secret | address was not verified in acceptInvitation, meaning that any account could 'steal' the invitation from the rightful owner if they front run a transaction.  Added test for this scenario and modified some tests after this additional verification.
The tryCatch function in exceptions.js wasn't actually doing anything, so all tests calling catchRevert were not acting as expected.  This highlighted a test that was actually failing even though it was previously being reported as passing; removeMember was not reverting if the account to remove was not already a member of the group.  Added a verification for this.